Turn elements into package
