+++ pyat/at/lattice/elements/__init__.py
@@ -1,0 +1,14 @@
+""":py:class:`.Element` objects used in AT.
+
+Each element has a default PassMethod attribute for which it should have the
+appropriate attributes.  If a different PassMethod is set, it is the caller's
+responsibility to ensure that the appropriate attributes are present.
+"""
+
+from .abstract_elements import *
+from .element_object import *
+from .basic_elements import *
+from .magnets import *
+from .crabcavity_element import *
+from .idtable_element import *
+from .variable_elements import *
